chore(prompt): slim default runtime guidance
